// onboarding/tests/contract/consumer/provider.enum.ts
export enum ProvidersEnum {
  Auth = 'nodejs-backend-auth',
  Commerceos = 'nodejs-backend-commerceos',
  Media = 'nodejs-backend-media',
  Users = 'nodejs-backend-users',
  Checkout = 'nodejs-backend-checkout',
  Wallpapers = 'nodejs-backend-wallpapers',
  Pos = 'nodejs-backend-pos',
  Connect = 'nodejs-backend-connect',
  Qr = 'connect-qr',
  DevicePayments = 'nodejs-backend-device-payments',
  CommunicationsThirdParty = 'third-party-communications',
  PaymentsThirdParty = 'third-party-payments',
}